3 Records with Maximum rating are retrived

+++ product.cs
@@ -10,7 +10,7 @@ namespace ProductReviewManagement
     {
         public int ProductID { get; set; }
         public int Userid { get; set; }
-        public int rating { get; set; }
+        public double rating { get; set; }
         public string Review { get; set; }
         public bool Islike { get; set; }
     }
